requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask_WTF==0.15.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.23.0
six==1.16.0
urllib3==1.25.9
Werkzeug==1.0.1
neomodel==4.0.5
python-dotenv==0.14.0